Switch `data` DataChannel to use JSON encoding rather than msgpack
Reduces the dependancy payload as were mostly just using the stdlib now.

Compartatively, this is 2x the transmit size at 3.3kb/frame, however this is perfectly acceptable for our uses.

+++ onboard/simulator.go
@@ -34,12 +34,12 @@ func (s *SimulatedSensor) GetValue(row, col int) uint8 {
 func (s *SimulatedSensor) GetState() (state SensorState) {
 	state = make(SensorState, s.rows)
 	for i := 0; i < s.rows; i++ {
-		state[i] = make([]uint8, s.cols)
+		state[i] = make([]int, s.cols)
 		for j := 0; j < s.cols; j++ {
 			if i == j {
 				state[i][j] = 0
 			} else {
-				state[i][j] = s.GetValue(i, j)
+				state[i][j] = int(s.GetValue(i, j))
 			}
 		}
 	}
